+++ packages/components/pages/f-contact-preferences/src/services/mapping/contactPreferences.mapper.js
@@ -7,11 +7,17 @@ import {
  * Maps the raw api response to the view model (hides and disables according to requirements).
  * @param {object} Preferences - The preferences contained in the passing object parameter (i.e. response.data.Preferences).
  * @param {object} PreferenceVersionViewed - The version value contained in the passing object parameter (i.e. response.data.PreferenceVersionViewed).
+ * @returns {object} A model that can be used to presents the view component.
  */
 export const mapToPreferencesViewModel = ({ Preferences, PreferenceVersionViewed }) => {
     const model = {
         preferences: Preferences.map(({
-            DisplayName, Sort, Key, Push, Email, Sms
+            DisplayName,
+            Sort,
+            Key,
+            Push,
+            Email,
+            Sms
         }) => ({
             displayName: DisplayName,
             sort: Sort,
@@ -34,11 +40,17 @@ export const mapToPreferencesViewModel = ({ Preferences, PreferenceVersionViewed
  * Maps the view model to the Body payload for the api update request.
  * @param {object} preferences - The preferences contained in the passing object parameter (i.e. state.preferences).
  * @param {object} preferenceVersionViewed - The version value contained in the passing object parameter (i.e. state.preferenceVersionViewed).
+ * @returns {object} A model that can be 'Posted' to the Api endpoint.
  */
 export const mapToPreferencesUpdateModel = ({ preferences, preferenceVersionViewed }) => {
     const model = {
         Preferences: preferences.map(({
-            displayName, sort, key, pushValue, emailValue, smsValue
+            displayName,
+            sort,
+            key,
+            pushValue,
+            emailValue,
+            smsValue
         }) => (
             {
                 DisplayName: displayName,
@@ -58,6 +70,11 @@ export const mapToPreferencesUpdateModel = ({ preferences, preferenceVersionView
     return model;
 };
 
+/**
+ * Filters out any preferences that are not marked as `visible` and sorts the array according to the `Sort` property.
+ * @param {object} preferences - The preferences (i.e. state.preferences).
+ * @returns {Array} Filtered and Sorted Preferences (If no preferences supplied then merely returns an empty array).
+ */
 export const filterSortPreferences = preferences => {
     const model = preferences?.filter(x => x.visible)?.sort((a, b) => a.sort - b.sort);